feat: edit distance
